MetsParser now uses an IMetsLoader for file ops
