+++ src/DigitalPreservation/XmlGen.Tests/MetsManagerWithPremis.cs
@@ -1,10 +1,7 @@
-﻿using Amazon.S3;
-using DigitalPreservation.Common.Model.Mets;
-using DigitalPreservation.Common.Model.Transit;
+﻿using DigitalPreservation.Common.Model.Transit;
 using FluentAssertions;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
-using Moq;
 using Storage.Repository.Common.Mets;
 using Storage.Repository.Common.Mets.StorageImpl;
 
@@ -23,8 +20,8 @@ public class MetsManagerWithPremis
 
         var factory = serviceProvider.GetService<ILoggerFactory>();
         var parserLogger = factory!.CreateLogger<MetsParser>();
-        var s3Client = new Mock<IAmazonS3>().Object;
-        parser = new MetsParser(s3Client, parserLogger);
+        var metsLoader = new FileSystemMetsLoader();
+        parser = new MetsParser(metsLoader, parserLogger);
         var metsStorage = new FileSystemMetsStorage(parser);
         metsManager = new MetsManager(parser, metsStorage);
     }
